+++ controllers/api/index.js
@@ -4,15 +4,11 @@ const muscleRoutes = require('./muscle-routes');
 const equipmentRoutes = require('./equipment-routes');
 const userRoutes = require('./equipment-routes');
 const routineRoutes = require('./routine-routes');
-// const auths = require('./auths')
-const withAuth = require('../../utils/auth')
 
-// router.use('/auth', auths)
 router.use('/exercise', exerciseRoutes);
 router.use('/muscle', muscleRoutes);
 router.use('/equipment', equipmentRoutes);
 router.use('/user', userRoutes);
 router.use('/routine', routineRoutes);
-
 
 module.exports = router;
