+++ controllers/api/routine-routes.js
@@ -1,9 +1,9 @@
 const router = require('express').Router();
 const { Routine, Exercise, Routine_Exercise } = require('../../models');
+// const withAuth = require('../../utils/auth');
 
 router.get('/', async (req, res) => {
   try {
-    console.log('blaablabla')
     const routineData = await Routine.findAll({
       include: [{model: Routine}],
     });
@@ -19,18 +19,15 @@ router.get('/:id', async (req, res) => {
     const routineData = await Routine_Exercise.findByPk(req.params.id, {
       include: [{model: Exercise}],
     });
-
     if (!routineData) {
       res.status(404).json({ message: 'No routine found with that id!' });
       return;
     }
-
     res.status(200).json(routineData);
   } catch (err) {
     res.status(500).json(err);
   }
 });
-
 
 router.post('/', async (req, res) => {
   try{
@@ -52,6 +49,5 @@ router.post('/:routine&:exercise', async (req, res) => {
     res.status(500).json(err);
   }
 });
-
 
 module.exports = router;
