+++ controllers/api/user-routes.js
@@ -13,8 +13,6 @@ router.get('/', async (req, res) => {
   }
 });
 
-
-
 router.get('/:id', async (req, res) => {
   try {
     const userData = await User.findByPk(req.params.id, {
