+++ controllers/html/home-page.js
@@ -1,0 +1,9 @@
+const router = require('express').Router();
+
+// Homepage
+router.get('/', (req, res) => {
+  res.render('homepage');
+});
+
+module.exports = router;
+
