+++ controllers/html/login-page.js
@@ -1,0 +1,9 @@
+const router = require('express').Router();
+
+router.get('/login', async (req, res, next) => {
+  res.render('login');
+  return;
+});
+
+module.exports = router;
+
