+++ models/Routine.js
@@ -11,7 +11,7 @@ Routine.init(
             primaryKey: true,
         },
         user_id: {      
-            type: DataTypes.UUID,
+            type: DataTypes.STRING,
             references: {
               model: 'user',
               key: 'id',
