+++ models/user.js
@@ -1,21 +1,37 @@
-const { Model, DataTypes, Sequelize, STRING } = require('sequelize');
+const { Model, DataTypes, Sequelize } = require('sequelize');
 // const { Sequelize } = require('sequelize/types')
 const sequelize = require('../config/connection')
 const bcrypt = require('bcrypt')
 
 class User extends Model {
-  async checkPassword(loginPw) {
-    return await bcrypt.compare(loginPw, this.password)
-  }
+  // async checkPassword(loginPw) {
+  //   return await bcrypt.compare(loginPw, this.password)
+  // }
 }
 
 User.init(
   {
     id: {
-      type: DataTypes.UUID,
+      type: DataTypes.STRING,
       defaultValue: Sequelize.UUIDV4,
       required: true,
       primaryKey: true,
+    },
+    displayName: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    firstName: {
+        type: DataTypes.STRING,
+        allowNull: false,
+    },
+    lastName: {
+        type: DataTypes.STRING,
+        allowNull: false,
+    },
+    image: {
+        type: DataTypes.STRING,
+        allowNull: true,
     },
     email: {
       type: DataTypes.STRING,
@@ -24,6 +40,7 @@ User.init(
     },
     password: {
       type: DataTypes.STRING,
+      defaultValue: Sequelize.UUIDV4,
       allowNull: false,
       validate: {
         len: [8],
