--- public/jham js copy/edit-dish.js
@@ -1,74 +1,0 @@
-async function editFormHandler(event) {
-    event.preventDefault();
-    const dish_name = document.querySelector('#dish_name').value;
-    const description = document.querySelector('#description').value;
-    const guest_name = document.querySelector('#guest_name').value;
-    
-    // What will the value of has_nuts be if the box in the form is checked? 
-    // The value of has_nuts will be true if the box is checked. 
-    // What do we call this kind of operator?
-    // We call this a ternary operator. It begins with a condition followed by a question mark and two code blocks separated by a :.
-    const has_nuts = document.querySelector('#has_nuts:checked') ? true : false;
-  
-  // window.location gives us access to the URL. We then use the .split() method to access the number at the end of the URL and set that equal to id.
-    const id = window.location.toString().split('/')[
-      window.location.toString().split('/').length - 1
-    ];
-  
-    // What part of our application will handle this 'put' request?
-    // The Controller will handle this 'put' request.
-  
-    const response = await fetch(`/api/dish/${id}`, {
-      method: 'PUT',
-      body: JSON.stringify({
-        dish_name,
-        description,
-        guest_name,
-        has_nuts,
-      }),
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    });
-  
-    // What happens if the response is ok?
-    // If the response is ok, that means that the dish was updated successfully. 
-    if (response.ok) {
-      document.location.replace(`/dish/${id}`);
-    } else {
-      alert('Failed to edit dish');
-    }
-  }
-  
-  document.querySelector('.edit-dish-form').addEventListener('submit', editFormHandler);
-
-
-//   this below goes into the controllers folder, it's just to see the front end above and here
-//   above corresponds with controllers > indexedDB.js
-  // According to MVC, what is the role of this action method?
-// This action method is the Controller. It accepts input and sends data to the Model and the View.
-router.put('/:id', async (req, res) => {
-    // Where is this action method sending the data from the body of the fetch request? Why?
-    // It is sending the data to the Model so that one dish can be updated with new data in the database.
-    try {
-      const dish = await Dish.update(
-      {
-        dish_name: req.body.dish_name,
-        description: req.body.description,
-        guest_name: req.body.guest_name,
-        has_nuts: req.body.has_nuts,
-      },
-      {
-        where: {
-          id: req.params.id,
-        },
-      });
-      // If the database is updated successfully, what happens to the updated data below?
-      // The updated data (dish) is then sent back to handler that dispatched the fetch request.
-      res.status(200).json(dish);
-    } catch (err) {
-        res.status(500).json(err);
-      };
-  });
-
-// How can I make this edit-routine.js code?
