--- public/login.html
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login-Stretch and Strengthen</title>
-    <!-- Google fonts -->
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Lato:wght@400;700;900&display=swap" rel="stylesheet">
-    <!-- Font Awesome -->
-	<script src="https://kit.fontawesome.com/45e97f819a.js" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.2/css/bulma.min.css">
-    <link rel="stylesheet" href="../public/css/style.css">
-    <!-- jQuery -->
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-</head>
-<body>
-   
-   <main>
-    <section id="login-section" class="hero is-fullheight">
-        <div class="hero-body">
-            <div class="container has-text-centered">
-                <div class="column is-4 is-offset-4">
-                    <h2 class="title">Login</h2>
-                    <hr class="login-hr">
-                    <p class="login-subtitle">Please login to proceed.</p>
-                    <div class="avatar-box">
-                        <figure class="avatar">
-                            <img src="./assets/images/logo-main.png">
-                        </figure>
-                        <form class="form login-form">
-                            <div class="field">
-                                <div class="control email-login">
-                                    <input class="input is-large" type="text" id="email-login" placeholder="Your Email" autofocus="">
-                                </div>
-                            </div>
-
-                            <div class="field">
-                                <div class="control password-login">
-                                    <input class="input is-large" type="password" id="password-login" placeholder="Your Password">
-                                </div>
-                            </div>
-                            <button class="button button-login-inverse is-block is-large is-fullwidth" id="loginbtn" type="submit">Login <i class="fa fa-sign-in" aria-hidden="true"></i></button>
-                        </form>
-                    </div>
-                    <a href ="/auth/google/callback" class="g-signin2" data-onsuccess="onSignIn"></a>
-                    <p>
-                        <a href="/signup">Sign Up</a>
-                    </p>
-                </div>
-            </div>
-        </div>
-    </section>
-   </main>
-</body>
-</html>
