--- public/signup.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Sign up-Stretch and Strengthen</title>
-    <!-- Google fonts -->
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Lato:wght@400;700;900&display=swap" rel="stylesheet">  
-    <!-- Font Awesome -->
-	<script src="https://kit.fontawesome.com/45e97f819a.js" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.2/css/bulma.min.css">
-    <link rel="stylesheet" href="../public/css/style.css">
-    <!-- jQuery -->
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-</head>
-<body>
-   
-   <main>
-    <section id="login-section" class="hero is-fullheight">
-        <div class="hero-body hero-signup">
-            <div class="container has-text-centered">
-                <div class="column is-4 is-offset-4 signup-block">
-                    <h2 class="title mt-3">Sign Up</h2>
-                    <hr class="signup-hr">
-                    <p class="signup-subtitle">Get on the ball and register.</p>
-                    <div class="avatar-box">
-                        <figure class="avatar">
-                            <img src="./assets/images/logo-main.png">
-                        </figure>
-                            <form class="form signup-form">
-                                <div class="field">
-                                    <div class="control name-signup">
-                                        <input class="input is-large" type="text" id="name-signup" placeholder=" Your Name" autofocus="">
-                                    </div>
-                                </div>
-
-
-                                <div class="field">
-                                    <div class="control email-signup">
-                                        <input class="input is-large" type="text" id="email-signup" placeholder="Your Email" autofocus="">
-                                    </div>
-                                </div>
-
-                                <div class="field">
-                                    <div class="control password-signup">
-                                        <input class="input is-large" type="password" id="password-signup" placeholder="Your Password">
-                                    </div>
-                                </div>
-
-                                <button class="button is-block button-signup-inverse is-large is-fullwidth" id="signupbtn" type="submit">Sign up </button>
-                            </form>
-                            <p class="mt-4"><a href="/login">Login</a></p>
-                    </div>
-                </div>
-                    
-            </div>
-        </div>
-
-    </section>
-   </main>
-</body>
-</html>
