+++ utils/auth.js
@@ -1,6 +1,6 @@
 const withAuth = (req, res, next) => {
     // If the user is not logged in, redirect the request to the login route
-    if (!req.session.logged_in) {
+    if (!req.session.loggedIn) {
       res.redirect('/login');
     } else {
       next();
