+++ views/about.handlebars
@@ -8,48 +8,45 @@
           </div>
       </div>
 </section>
+
 <section id="about-exercise-details" class="section exercise-details">
-    <div class="container">
-        {{!-- <h2 class="title is-size-2">Sample Routine</h2> --}}
-          {{!-- Large screen --}}
-          <div class="columns is-mobile is-centered is-hidden-touch">
-              {{!-- Column 50% --}}
-            <div class="column is-half">
-
-              {{!-- Nested column --}}
-                <div class="columns is-mobile is-multiline">
-                    <div class="column is-narrow-desktop">
-                        <figure>
-                          <div class="muscle-background mx-auto" style="background-image: url('https://wger.de/static/images/muscles/main/muscle-1.svg'),
-  url('https://wger.de/static/images/muscles/muscular_system_front.svg');">
-                          </div>
-                        </figure>
-                    </div>
-                    <div class="column">
-                      <h2 class="title is-size-2">Biceps Curls With Barbell</h2>
-                      <p>Hold the Barbell shoulder-wide, the back is straight, the shoulders slightly back, the arms are streched. Bend the arms, bringing the weight up, with a fast movement. Without pausing, let down the bar with a slow and controlled movement.</p>
-                      <p>Don't allow your body to swing during the exercise, all work is done by the biceps, which are the only mucles that should move (pay attention to the elbows).</p>
-                    </div>
-                 </div>
-
-            </div>
+  <div class="container">
+    {{!-- <h2 class="title is-size-2">Sample Routine</h2> --}}
+    {{!-- Large screen --}}
+    <div class="columns is-mobile is-centered is-hidden-touch">
+      {{!-- Column 50% --}}
+      <div class="column is-half">
+        {{!-- Nested column --}}
+        <div class="columns is-mobile is-multiline">
+          <div class="column is-narrow-desktop">
+            <figure>
+              <div class="muscle-background mx-auto" style="background-image: url('https://wger.de/static/images/muscles/main/muscle-1.svg'), url('https://wger.de/static/images/muscles/muscular_system_front.svg');"></div>
+            </figure>
           </div>
-          {{!-- Mobile and Tablet --}}
-          <div class="columns is-tablet is-hidden-desktop">
-            <div class="column">
-                <figure>
-                      <div class="muscle-background mx-auto" style="background-image: url('https://wger.de/static/images/muscles/main/muscle-1.svg'),
-url('https://wger.de/static/images/muscles/muscular_system_front.svg');">
-                      </div>
-                </figure>
-            </div>
-            <div class="column">
-              <h2 class="title is-size-2">Biceps Curls With Barbell</h2>
-                      <p>Hold the Barbell shoulder-wide, the back is straight, the shoulders slightly back, the arms are streched. Bend the arms, bringing the weight up, with a fast movement. Without pausing, let down the bar with a slow and controlled movement.</p>
-                      <p>Don't allow your body to swing during the exercise, all work is done by the biceps, which are the only mucles that should move (pay attention to the elbows).</p>
-            </div>
+
+          <div class="column">
+            <h2 class="title is-size-2">Biceps Curls With Barbell</h2>
+            <p>Hold the Barbell shoulder-wide, the back is straight, the shoulders slightly back, the arms are streched. Bend the arms, bringing the weight up, with a fast movement. Without pausing, let down the bar with a slow and controlled movement.</p>
+            <p>Don't allow your body to swing during the exercise, all work is done by the biceps, which are the only mucles that should move (pay attention to the elbows).</p>
           </div>
+        </div>
+      </div>
     </div>
+
+      {{!-- Mobile and Tablet --}}
+    <div class="columns is-tablet is-hidden-desktop">
+      <div class="column">
+        <figure>
+          <div class="muscle-background mx-auto" style="background-image: url('https://wger.de/static/images/muscles/main/muscle-1.svg'), url('https://wger.de/static/images/muscles/muscular_system_front.svg');"></div>
+        </figure>
+      </div>
+      <div class="column">
+        <h2 class="title is-size-2">Biceps Curls With Barbell</h2>
+          <p>Hold the Barbell shoulder-wide, the back is straight, the shoulders slightly back, the arms are streched. Bend the arms, bringing the weight up, with a fast movement. Without pausing, let down the bar with a slow and controlled movement.</p>
+          <p>Don't allow your body to swing during the exercise, all work is done by the biceps, which are the only mucles that should move (pay attention to the elbows).</p>
+      </div>
+    </div>
+  </div>
 </section>
