+++ views/login.handlebars
@@ -13,28 +13,31 @@
                         <form class="form login-form">
                             <div class="field">
                                 <div class="control email-login">
-                                    <input class="input is-large" type="text" id="email-login" placeholder="Your Email" autofocus="">
+                                    <input class="input is-large" type="text" id="email-login" placeholder="Your Email"
+                                        autofocus="">
                                 </div>
                             </div>
 
                             <div class="field">
                                 <div class="control password-login">
-                                    <input class="input is-large" type="password" id="password-login" placeholder="Your Password">
+                                    <input class="input is-large" type="password" id="password-login"
+                                        placeholder="Your Password">
                                 </div>
                             </div>
-                            <button class="button button-login-inverse is-block is-large is-fullwidth" id="loginbtn" type="submit">Login <i class="fa fa-sign-in" aria-hidden="true"></i></button>
+                            <button class="button button-login-inverse is-block is-large is-fullwidth" id="loginbtn"
+                                type="submit">Login <i class="fa fa-sign-in" aria-hidden="true"></i></button>
                         </form>
                     </div>
                     <!-- <div class="g-signin2" data-onsuccess="onSignIn"></div> -->
                     <p>
-                       <a href="/auth/google">Sign In with Google</a> <br>
+                        <a href="/auth/google">Sign In with Google</a> <br>
                         <a href="/signup">Sign Up</a>
                     </p>
                 </div>
             </div>
         </div>
     </section>
-   </main>
+</main>
 
 <script src="https://apis.google.com/js/platform.js" async defer></script>
 <script src="/js/login.js"></script>
