+++ views/muscle-specific.handlebars
@@ -5,7 +5,7 @@
               <div class="column column-custom-center">
                   <ul class="ss-list-group">
                     {{#each muscle.exercises as |exercise|}}
-                    {{> exercise-details}}
+                      {{> exercise-details}}
                     {{/each}}
                   </ul>
               </div>
