+++ views/muscles.handlebars
@@ -6,7 +6,7 @@
                  <h1 class="title is-size-1 has-text-centered">Muscles</h1>
                      <ul class="ss-list-group">
                         {{#each muscles as |muscle|}}
-                        {{> muscles-details}}
+                            {{> muscles-details}}
                         {{/each}}
                     </ul>
             </div>
