+++ views/signup.handlebars
@@ -10,36 +10,39 @@
                         <figure class="avatar">
                             <img src="./assets/images/logo-main.png">
                         </figure>
-                            <form class="form signup-form">
+                        <form class="form signup-form">
 
-                                <div class="field">
-                                    <div class="control email-signup">
-                                        <input class="input is-large" type="text" id="email-signup" placeholder="Your Email" autofocus="">
-                                    </div>
+                            <div class="field">
+                                <div class="control email-signup">
+                                    <input class="input is-large" type="text" id="email-signup" placeholder="Your Email"
+                                        autofocus="">
                                 </div>
+                            </div>
 
-                                <div class="field">
-                                    <div class="control password-signup">
-                                        <input class="input is-large" type="password" id="password-signup" placeholder="Your Password">
-                                    </div>
+                            <div class="field">
+                                <div class="control password-signup">
+                                    <input class="input is-large" type="password" id="password-signup"
+                                        placeholder="Your Password">
                                 </div>
+                            </div>
 
-                                <button class="button is-block button-signup-inverse is-large is-fullwidth" id="signupbtn" type="submit">Sign up </button>
-                            </form>
-                            
-                            <p class="mt-4">
-                                <a href="/auth/google">Sign Up with Google</a> <br>
-                                <a href="/login">Login</a>
-                                </p>
+                            <button class="button is-block button-signup-inverse is-large is-fullwidth" id="signupbtn"
+                                type="submit">Sign up </button>
+                        </form>
+
+                        <p class="mt-4">
+                            <a href="/auth/google">Sign Up with Google</a> <br>
+                            <a href="/login">Login</a>
+                        </p>
                     </div>
                 </div>
-                    
+
             </div>
         </div>
 
     </section>
-   </main>
+</main>
 
-    <!-- javascript file-->
+<!-- javascript file-->
 <script src="https://apis.google.com/js/platform.js" async defer></script>
 <script src="../js/login.js"></script>
